Change to using the official PNG files.
Rewrite tests and add new files tIME.t and tEXT.t for testing.

Remove old time and text testing from libpng.t. That relied on a file
which was actually broken, causing problems at cpantesters on some
machines (the iTXT compression was borked, set to 0 when this is not a
valid value for iTXT comopresssion, and was getting 1 and 2 dependong
on the computer).