Update Libpng.xs.tmpl
Using `<png.h>` instead of `"png.h"` is causing issues when trying to build this on FreeBSD (using clang, not gcc).  FreeBSD enforces stricter separation between "system library" (what's in /usr/include) and "userland" (what's in /usr/local/include) than Linux does...  Since png.h is not part of the Base System installation, it needs to be doublequoted to be found.

Another way to do this would be to add `-I /usr/local/include` to CFLAGS or INC somewhere, but this is what worked for me.